requirements.txt
replicate==0.11.0
openai==0.27.9
requests==2.31.0
nltk==3.8.1
beautifulsoup4==4.12.2
Pillow==10.0.0
Markdown==3.4.4
tqdm==4.66.1
